Protocol: Augment OpenChannel to also contain STmin.
STmin is the minimum separation time between consecutive ISOTP frames.
In `OpenChannel`, it is encoded as the third parameter w/ RX in the high
nibble and TX in the low nibble.

Note that for TX, our semantics for STmin is not plainly adding this to the
received value in the flow control, but rather taking the maximum value.