libCANyonero: Specify on-the-wire format for RPC commands
RPC commands are for exchanging JSON-formatted payloads between the central and the peripheral.
The format is loosely based on JSON-RPC. An example call to gather the number of stored coredumps
may look like that:
```json
{
  "id": 0,
  "method": "system.coredumps.list",
  "params": null
}
```

A possible answer could be:

```json
{
  "id": 0,
  "result": [
      "coredump-0.5.24-22.bin",
      "coredump-0.5.24-26.bin",
      "coredump-0.5.25-30.bin"
   ]
}
```

RPC commands are intended to perform maintenance and/or diagnostic functions without
having to continually enhance the CANyonero wire protocol.
They're not intended to perform time critical operations or superseed the core
functionality.